factor MultiDict into server utils
