+++ bokeh/server/utils/multi_dict.py
@@ -1,0 +1,26 @@
+#-----------------------------------------------------------------------------
+# Copyright (c) 2012 - 2015, Continuum Analytics, Inc. All rights reserved.
+#
+# Powered by the Bokeh Development Team.
+#
+# The full license is in the file LICENSE.txt, distributed with this software.
+#-----------------------------------------------------------------------------
+
+class MultiDict(object):
+
+    def __init__(self):
+        self.dict = {}
+
+    def add(self, k, v):
+        self.dict.setdefault(k, set()).add(v)
+
+    def remove_val(self, k, v):
+        self.dict.setdefault(k, set()).remove(v)
+        if len(self.dict[k]) == 0:
+            self.remove(k)
+
+    def remove(self, k):
+        del self.dict[k]
+
+    def get(self, *args):
+        return self.dict.get(*args)
